Fix bug that amqp_userinfo part in service URI is not recognized.
When URI with userinfo was set to Service URI, Invalid URL error was
raised in spite of fully following amqp uri spec. It was validator bug.

In addition, there is an issue regarding password in userinfo.
It is not blinded.

So, this patch isolates userinfo from Service URI then add dedicated
field to global configuration.